+++ src/Fxml/DeleteCourseRoom.fxml
@@ -10,7 +10,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<VBox prefHeight="700.0" prefWidth="1240.0" xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Fxml.DelDCController">
+<VBox prefHeight="700.0" prefWidth="1240.0" xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controller.DelDCController">
   <children>
     <AnchorPane maxHeight="-1.0" maxWidth="-1.0" prefHeight="-1.0" prefWidth="-1.0" VBox.vgrow="ALWAYS">
          <children>
@@ -26,13 +26,13 @@
           </MenuBar>
             <TableView fx:id="dctable" layoutX="179.0" layoutY="150.0" prefHeight="461.0" prefWidth="548.0" AnchorPane.leftAnchor="179.0" AnchorPane.topAnchor="150.0">
               <columns>
-                <TableColumn prefWidth="62.0" text="CheckBox" />
+                <TableColumn fx:id="dccheckColumn" prefWidth="62.0" text="CheckBox" />
                   <TableColumn fx:id="dcnameColumn" prefWidth="426.0" text="学科・コース" />
               </columns>
             </TableView>
             <TableView fx:id="crtable" layoutX="715.0" layoutY="150.0" prefHeight="461.0" prefWidth="250.0" AnchorPane.rightAnchor="200.0" AnchorPane.topAnchor="150.0">
               <columns>
-                <TableColumn prefWidth="60.0" text="CheckBox" />
+                <TableColumn fx:id="crcheckColumn" prefWidth="60.0" text="CheckBox" />
                   <TableColumn fx:id="crnameColumn" prefWidth="189.0" text="教室" />
               </columns>
             </TableView>
