+++ src/controller/DelDCController.java
@@ -1,4 +1,4 @@
-package Fxml;
+package controller;
 
 import java.net.URL;
 import java.util.ResourceBundle;
@@ -11,6 +11,7 @@ import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
+import javafx.scene.control.cell.CheckBoxTableCell;
 import javafx.scene.control.cell.PropertyValueFactory;
 
 public class DelDCController implements Initializable {
@@ -23,6 +24,10 @@ public class DelDCController implements Initializable {
 	private TableColumn<?, ?> dcnameColumn;
 	@FXML
 	private TableColumn<?, ?> crnameColumn;
+	@FXML
+	private TableColumn<DepartmentCourse, Boolean> dccheckColumn;
+	@FXML
+	private TableColumn<ClassRoom, Boolean> crcheckColumn;
 
 	@Override
 	public void initialize(URL location, ResourceBundle resources){
@@ -30,6 +35,10 @@ public class DelDCController implements Initializable {
 		crtable.setItems(ClassRoomDAO.selectDAO());
 		dcnameColumn.setCellValueFactory(new PropertyValueFactory<>("dcname"));
 		crnameColumn.setCellValueFactory(new PropertyValueFactory<>("crname"));
+		dccheckColumn.setCellFactory(CheckBoxTableCell.forTableColumn(dccheckColumn));
+		dccheckColumn.setCellValueFactory(new PropertyValueFactory<DepartmentCourse, Boolean>("dccheck"));
+		crcheckColumn.setCellFactory(CheckBoxTableCell.forTableColumn(crcheckColumn));
+		crcheckColumn.setCellValueFactory(new PropertyValueFactory<ClassRoom, Boolean>("crcheck"));
 	}
